Complete query-deduplication: add dedup utilities and tests

+++ internal/filter/bitmap.go
@@ -758,7 +758,7 @@ func normalizeInt64(value any) (int64, bool) {
 	case int64:
 		return v, true
 	case uint:
-		if v > uint64(maxInt64) {
+		if uint64(v) > uint64(maxInt64) {
 			return 0, false
 		}
 		return int64(v), true
